[Devesh]: Added ability to calculate total fare , Average fare and total rides for multiple rides

+++ src/main/java/com/cabinvoicegenerator/InvoiceData.java
@@ -3,9 +3,9 @@ package com.cabinvoicegenerator;
 public class InvoiceData {
 	
 	public Double kiloMeters;
-	public Integer time;
+	public Double time;
 	
-	public InvoiceData(Double kiloMeters,Integer time) {
+	public InvoiceData(Double kiloMeters,Double time) {
 		this.kiloMeters = kiloMeters;
 		this.time = time;
 	}
